requirements.txt
datasets==2.18.0
gin-config==0.5.0
ipykernel==6.29.4
matplotlib==3.8.4
numpy==1.26.4
plotly==5.22.0
nbformat>=4.2.0
torch==2.3.0
pytorch-lightning==2.2.4
torchmetrics==1.4.0
tensorboard==2.16.2